WIP GM:S room rendering

+++ src/KEngine/include/KEngine/Entity/Components/EntityRenderableComponents.hpp
@@ -9,6 +9,7 @@
 #include "KEngine/Common/Point2D.hpp"
 
 #include <cassert>
+#include <cstdint>
 
 namespace ke
 {
@@ -53,7 +54,7 @@ namespace ke
         KE_DEFINE_ENTITY_COMPONENT_COMMON_PROPERTIES(EntityRenderableCircleShapeComponent, 0x8617353F)
 
     public:
-        EntityRenderableCircleShapeComponent(ke::EntitySptr entity, const ke::Transform2D & localTransform, std::int16_t depth,
+        EntityRenderableCircleShapeComponent(ke::EntitySptr entity, const ke::Transform2D & localTransform, std::int32_t depth,
             const ke::Color & fillColor, float radius = 5.0f, const ke::Color & outlineColor = ke::Color::TRANSPARENT,
             float outlineThickness = 0.0f)
             : EntityRenderableComponent(entity)
@@ -69,7 +70,7 @@ namespace ke
         KE_DEFINE_ENTITY_COMPONENT_COMMON_PROPERTIES(SpriteDrawableComponent, 0xdc84005a)
 
     public:
-        SpriteDrawableComponent(ke::EntitySptr entity, const ke::Transform2D & localTransform, std::int16_t depth, size_t textureId,
+        SpriteDrawableComponent(ke::EntitySptr entity, const ke::Transform2D & localTransform, std::int32_t depth, size_t textureId,
             const ke::Rect2DInt32 & textureRect, const ke::Color & color = ke::Color::WHITE)
             : EntityRenderableComponent(entity)
         {
@@ -84,7 +85,7 @@ namespace ke
 
     public:
         LineDrawableComponent(ke::EntitySptr entity, const Point2DFloat & begin, const Point2DFloat & end,
-            std::int16_t depth, const ke::Color & color = ke::Color::WHITE)
+            std::int32_t depth, const ke::Color & color = ke::Color::WHITE)
             : EntityRenderableComponent(entity)
         {
             this->sceneNode = ke::LineNode::create(entity->getId(), begin, end, depth, color);
